net: phy: mv88e61xx: add 6390 support
On the whole, this chip is very similar to its predecessors. One major
difference is that the PHYs are no longer accessible in the SMI
address space, only via indirect access. Because of this, port
registers have been moved down to offset 0. Additionally, the CPU port
is now port 0, and the first external copper port is port 1 etc.