net: phy: mv88e61xx: add mdio bus for top-level smi
Previously the "indirect smi" bus was registered with the system to
allow PHY register read/write by generic interfaces like the `mii`
command. But there was no interface to read the switch registers.

Add a new mdio bus called "smi" which can be used to read switch
registers. For consistency, rename the existing bus to "smi-phy" to
make the distinction clearer. This name is also easier to use from the
shell without running in to quoting issues.