spl: add config option to send environment params to itb script
Some of these scripts, e.g. the imx8 flavor, takes parameters in the
form of environment variables. Existing READMEs mandate these be set
by the user at build time. This violates the typical 2-step process of
configure and build.

With this option, the parameters can be stored in the _defconfigs
instead.